37강 애니메이션 bmp 분할 설계 -> 애니메이션을 우선 한 줄 분할을 해서 frm 으로 애니메이션 내부에서 저장을 하는 것 까지 했다 + 구조적인 문제점 인 ResourceMgr 을 손봤다. ( Texture 생성 , 소멸자 접근에 관한 문제 )

+++ WindowsProject1/CCollider.cpp
@@ -57,10 +57,10 @@ void CCollider::render(HDC _dc)
 	// 나중에 소멸하니까 소멸자를 무조건 호출하게 되어 있다.
 
 	Rectangle(_dc
-		, m_vFinalPos.x - m_vScale.x / 2.f
-		, m_vFinalPos.y - m_vScale.y / 2.f
-		, m_vFinalPos.x + m_vScale.x / 2.f
-		, m_vFinalPos.y + m_vScale.y / 2.f);
+		, (float)(m_vFinalPos.x - m_vScale.x / 2.f)
+		, (float)(m_vFinalPos.y - m_vScale.y / 2.f)
+		, (float)(m_vFinalPos.x + m_vScale.x / 2.f)
+		, (float)(m_vFinalPos.y + m_vScale.y / 2.f));
 
 	//SelectObject(_dc, hDefaultBrush);
 	//SelectObject(_dc, hDefaultPen);
